requirements.txt
python-dotenv==1.0.1
requests==2.31.0
spotipy==2.23.0
beautifulsoup4==4.12.2